secured product routes and modified is isUserMiddleware to isSameUserMiddleware
